+++ frontend/src/metabase/visualizations/components/TableInteractive/hooks/use-object-detail.ts
