+++ frontend/src/metabase/visualizations/components/TableInteractive/index.ts
