+++ frontend/src/metabase/visualizations/components/TableRoot/TableRoot.module.css
@@ -1,0 +1,7 @@
+.TableID .cellData {
+  border: 1px solid transparent;
+  padding: 0.25em 0.65em;
+  border-radius: 99px;
+  min-width: 35px;
+  text-align: center;
+}
